pyproject.toml
[tool.poetry]
name = "ts-drf"
version = "0.1.0"
description = ""
authors = ["matheus <matheus.tessaroli@olist.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
